file name change to user_authenticate.js, bearer token added
